Restructure of Swappable
