+++ src/Swappable/index.js
@@ -1,0 +1,3 @@
+import Swappable from './Swappable';
+
+export default Swappable;
