+++ src/index.js
@@ -1,7 +1,7 @@
 import Draggable from './Draggable';
 import Droppable from './Droppable';
+import Swappable from './Swappable';
 import Sortable from './sortable';
-import Swappable from './swappable';
 import AbstractEvent from './events/abstract-event';
 import Snappable from './behaviour/snappable';
 import Collidable from './behaviour/collidable';
